+++ locations/src/main/kotlin/io/bkbn/kompendium/locations/NotarizedLocations.kt
@@ -51,17 +51,17 @@ object NotarizedLocations {
     val spec = application.attributes[KompendiumAttributes.openApiSpec]
     val serializableReader = application.attributes[KompendiumAttributes.schemaConfigurator]
     pluginConfig.locations.forEach { (k, v) ->
-      val path = Path()
-      path.parameters = v.parameters
-      v.get?.addToSpec(path, spec, v, serializableReader)
-      v.delete?.addToSpec(path, spec, v, serializableReader)
-      v.head?.addToSpec(path, spec, v, serializableReader)
-      v.options?.addToSpec(path, spec, v, serializableReader)
-      v.post?.addToSpec(path, spec, v, serializableReader)
-      v.put?.addToSpec(path, spec, v, serializableReader)
-      v.patch?.addToSpec(path, spec, v, serializableReader)
-
       val location = k.getLocationFromClass()
+      val path = spec.paths[location] ?: Path()
+      path.parameters = path.parameters?.plus(v.parameters) ?: v.parameters
+      v.get?.addToSpec(path, spec, v, serializableReader, location)
+      v.delete?.addToSpec(path, spec, v, serializableReader, location)
+      v.head?.addToSpec(path, spec, v, serializableReader, location)
+      v.options?.addToSpec(path, spec, v, serializableReader, location)
+      v.post?.addToSpec(path, spec, v, serializableReader, location)
+      v.put?.addToSpec(path, spec, v, serializableReader, location)
+      v.patch?.addToSpec(path, spec, v, serializableReader, location)
+
       spec.paths[location] = path
     }
   }
